Conflicts:
	InstentMusic/.settings/org.eclipse.wst.common.component
	InstentMusic/src/main/webapp/WEB-INF/views/chattingTemp.jsp
	InstentMusic/src/main/webapp/WEB-INF/views/main.jsp
	InstentMusic/src/main/webapp/WEB-INF/views/music/makingMusic.jsp
	InstentMusic/target/classes/global/sesoc/teamBOB4/HomeController.class
	InstentMusic/target/m2e-wtp/web-resources/META-INF/maven/global.sesoc/teamBOB4/pom.properties

+++ InstentMusic/src/main/java/global/sesoc/teamBOB4/MusicController.java
@@ -16,7 +16,9 @@ import org.springframework.web.multipart.MultipartFile;
 
 import global.sesoc.teamBOB4.dao.MusicDao;
 import global.sesoc.teamBOB4.util.FileService;
+import global.sesoc.teamBOB4.vo.Key_sound;
 import global.sesoc.teamBOB4.vo.Sound_library;
+import global.sesoc.teamBOB4.vo.Temp;
 
 @Controller
 @RestController
@@ -40,6 +42,16 @@ public class MusicController {
 		List<Sound_library> result = dao.getSounds(sound);
 		for(Sound_library s : result) {
 			String fullPath = "resources/"+uploadPath+s.getSou_saved();
+			if(sound.getSou_type().equals("added")) {
+				Sound_library added = new Sound_library();
+				int a = Integer.parseInt(s.getSou_name());
+				added.setSou_number(a);
+				List<Sound_library> original = dao.getorigin(added);
+				for(Sound_library o : original) {
+					fullPath = "resources/"+uploadPath+o.getSou_saved();
+					s.setSou_name(o.getSou_name());
+				}
+			}
 			s.setFullPath(fullPath);
 		}
 		return result;
@@ -77,16 +89,18 @@ public class MusicController {
 	
 	@PostMapping("/deletelib")
 	public int deletelib(Sound_library sound, HttpSession session, HttpServletRequest request) {
-		String rootPath = request.getSession().getServletContext().getRealPath("/") ;//리얼경로
-		String savePath = rootPath + "/resources/"+uploadPath ;
-		
 		sound.setCust_number((int)session.getAttribute("login"));
-		List<Sound_library> result = dao.getSounds(sound);
-		
-		for(Sound_library r : result) {
-			String fullPath = savePath+r.getSou_saved();
+		if(!sound.getSou_type().equals("added")) {
+			String rootPath = request.getSession().getServletContext().getRealPath("/") ;//리얼경로
+			String savePath = rootPath + "/resources/"+uploadPath ;
 			
-			FileService.deleteFile(fullPath);
+			List<Sound_library> result = dao.getSounds(sound);
+			
+			for(Sound_library r : result) {
+				String fullPath = savePath+r.getSou_saved();
+				
+				FileService.deleteFile(fullPath);
+			}
 		}
 		return dao.deletelib(sound);
 	}
@@ -110,5 +124,29 @@ public class MusicController {
 		} else {
 			return "fail";
 		}
+	}
+	
+	@PostMapping("/insertkey")
+	public int insertkey(Key_sound keys, HttpSession session) {
+		int cust = (int) session.getAttribute("login");
+		keys.setCust_number(cust);
+		return dao.insertkey(keys);
+	}
+	
+	@PostMapping("/delkey")
+	public int delkey(Key_sound keys) {
+		return dao.delkey(keys);
+	}
+	
+	@PostMapping("/updatekey")
+	public int updatekey(Key_sound keys) {
+		return dao.updatekey(keys);
+	}
+	
+	@GetMapping("/getkeys")
+	public List<Key_sound> getkeys(Key_sound keys, HttpSession session){
+		int cust = (int) session.getAttribute("login");
+		keys.setCust_number(cust);
+		return dao.getkeys(keys);
 	}
 }
